pyproject.toml
[tool.poetry]
name = "garagemav"
version = "0.1.0"
description = ""
authors = ["Amanda Vincenzi <amandavincenzi.ifc@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1"
gunicorn = "^20.1.0"
whitenoise = "^6.2.0"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
